Speed up Augeas resources by specifying the lens.
Augeas resources are extremely slow when using a context instead of
a specific lens because Augeas has to compile all of the lenses to
decide which lens to use. If the lens needed is already known,
it should be specified.

On an Ubuntu 12.04 box, Augeas usage per resource went from ~.55s
to .01s when measured using puppet-profiler.

This change also introduces a new php::params variable
$augeas_contrib_dir. The Augeas contrib dir was being repeated five
times in two separate files.