php/init.php: do not pass $ensure to extensions
most of the time this is okay for PHP extensions, because these are
just packaged differently because of the distro and they will have
the PHP core version number (i.e. soap will be 5.5.9-something).

some extensions do proper versioning. in this case we do not want
to pass this.