Pipe was broken in the previous rubucop fix
Pipe was being set previously, 2 commits back, but then not used so rubucop complained.  1 commit back, pipe was then changed to being compared, which broke it because it was never set.  This fixes that to remove the variable `pipe` altogether since we only use @resource[:pipe] once and don't need a variable for it.  Brings it in line with things like @resource[:source]