use voxpupuli/archive to download composer
rather than using a hand-crafted curl call, we use voxpupuli/archive to
download composer. This has the benefit that it doesn't require us to
$manage_curl. But we also remove `$environment`! This variable name is
poorly chosen. Overriding it (or even just setting it to `undef`), could
mean that someone's `hiera()` lookups are failing, as soon as they enter
our module!

This pull-request addresses #273.